(Test Harness) The test harness prints exit status
Normally, when a command is run from a makefile it is done so using
a shell (such as /bin/sh) and if the command dies due to a signal
the shell prints a message to that effect.  On the other hand, when
a command is run from the test harness it isn't always run through
a shell and thus the signal-death message isn't always printed.

This commit changes the test harness so it always prints the
signal-death message, and it also even prints whenever a command
has a non-zero exit status.

Also replaced TAB characters with SPC.