(Lexer) Fixed bug when doing lexical analysis for C strings
An extraneous rule for double-quote prevented the lexer from
entering the STRING_LIT state.  If the string contained what
looked like the start of a C comment, the lexer inappropriately
entered the C_COMMENT state.

This commit removes the extraneous double-quote rule.  It also
removes string-related commentary that is no longer applicable.